Incorporate sonar scanner
Currently, this is failing on some of the
sonarcloud.io connections when I run it locally
e.g. fails to download a plugin or connect to
the server. Going to chance it that it's just
a local network issue on my VM and not something
coded incorrectly.

Will update action before pushing